Ahora mismo hay un almacen para los revocations temporales pero no para los finales (arreglando código por problemas anteriores)

+++ src/main/python/secure_all/data/revocation.py
@@ -5,6 +5,8 @@ import json
 from secure_all.exception.access_management_exception import AccessManagementException
 from secure_all.data.access_key import AccessKey
 from secure_all.storage.keys_json_store import KeysJsonStore
+from secure_all.storage.revocations_json_store import RevocationsJsonStore
+from secure_all import JSON_FILES_PATH
 
 # pylint:disable=invalid-name
 # pylint:disable=no-self-use
@@ -28,7 +30,11 @@ class Revocation:
             except json.JSONDecodeError as ex:
                 raise AccessManagementException("JSON Decode Error - Wrong JSON Format") from ex
 
-            key = data["Key"]
+            try:
+                key = data["Key"]
+            except KeyError as ex:
+                raise AccessManagementException("La clave recibida no existe") from ex
+
             store = KeysJsonStore()
             try:
                 found_key = AccessKey.create_key_from_id(key)
@@ -40,15 +46,38 @@ class Revocation:
 
             if found_key is None:
                 raise AccessManagementException("La clave recibida no existe")
-
+            """
+            print(found_key.revocation)
             if found_key.revocation:
                 raise AccessManagementException("La clave fue revocada previamente por este método")
             found_key.revocation = True
+            print(found_key.revocation)
+            """
 
             emails = found_key.notification_emails
-            store.delete_item(found_key.key)
+
+            if data["Revocation"] == "Final":
+                store.delete_item(found_key.key)
+                store.save_store()
+
+            elif data["Revocation"] == "Temporal":
+                store_revocations = RevocationsJsonStore()
+                if store_revocations.find_item(found_key.key) is None:
+                    store.delete_item(found_key.key)
+                    store.save_store()
+
+                    store_revocations.add_item(found_key)
+                    store_revocations.save_store()
+                else:
+                    raise AccessManagementException("La clave fue revocada previamente por este método")
+
+            else:
+                raise AccessManagementException("El archivo de entrada tiene algún problema relacionado con su formato o con su acceso.")
+            """
             store.add_item(found_key)
             store.save_store()
+            print(store.__dict__)
+            """
             return emails
 
     __instance = None
